Started code splitting into helper
